sidecars/voice/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Audio processing
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1

# VAD
torch==2.1.2
torchaudio==2.1.2

# Speaker identification
resemblyzer==0.1.1.dev0

# Transcription
faster-whisper==0.10.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
